refactor(backup): refactor S3 backup to use service object and improve Sentry tests
- Extract S3 backup logic from S3BackupJob into a new S3BackupService class
- Simplify S3BackupJob to delegate to S3BackupService and log results
- Add S3Helpers concern to share S3 validation and error handling methods
- Introduce SentryTestJob and SentryTestService for improved Sentry integration testing
- Update rake tasks for S3 backup and Sentry to support immediate execution and job queue
- Add MissionControlController for admin authentication and update routes and config

This refactor improves code organization, testability, and observability of backup and error reporting features.
