+++ app/controllers/mission_control_controller.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class MissionControlController < ApplicationController
+  before_action :require_admin
+end
