--- lib/admin_constraint.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-class AdminConstraint
-  def matches?(request)
-    return false unless request.session[:user_id]
-    
-    user = User.find_by(id: request.session[:user_id])
-    user&.admin?
-  end
-end
